requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Werkzeug==2.3.6
bcrypt==4.0.1
email-validator==2.0.0
python-dotenv==1.0.0
python-magic==0.4.27